+++ src/controllers/AdminCmsLangController.php
@@ -47,7 +47,8 @@ class AdminCmsLangController extends BackendModelStandartController
                 ],
                 'grid'    => [
                     'defaultOrder' => [
-                        'active' => SORT_DESC,
+                        'is_default' => SORT_DESC,
+                        'is_active' => SORT_DESC,
                         'priority' => SORT_ASC,
                     ],
                     'visibleColumns' => [
@@ -57,11 +58,15 @@ class AdminCmsLangController extends BackendModelStandartController
                         'image_id',
                         'name',
                         'code',
-                        'active',
+                        'is_active',
+                        'is_default',
                         'priority',
                     ],
                     'columns'        => [
-                        'active'   => [
+                        'is_active'   => [
+                            'class' => BooleanColumn::class,
+                        ],
+                        'is_default'   => [
                             'class' => BooleanColumn::class,
                         ],
                         'image_id' => [
@@ -99,10 +104,11 @@ class AdminCmsLangController extends BackendModelStandartController
                 ],
             ],
             'code',
-            'active'   => [
+            'is_active'   => [
                 'class'      => BoolField::class,
-                'trueValue'  => "Y",
-                'falseValue' => "N",
+            ],
+            'is_default'   => [
+                'class'      => BoolField::class,
             ],
             'name',
             'description',