+++ src/migrations/m190412_185515__update_data_table__cms_lang.php
@@ -1,0 +1,24 @@
+<?php
+/**
+ * @author Semenov Alexander <semenov@skeeks.com>
+ * @link http://skeeks.com/
+ * @copyright 2010 SkeekS (СкикС)
+ * @date 28.08.2015
+ */
+
+use yii\db\Schema;
+use yii\db\Migration;
+
+class m190412_185515__update_data_table__cms_lang extends Migration
+{
+    public function safeUp()
+    {
+        $this->update("{{%cms_lang}}", ['is_default' => 1], ['def' => 'Y']);
+    }
+
+    public function safeDown()
+    {
+        echo "m190412_185515__update_data_table__cms_lang cannot be reverted.\n";
+        return false;
+    }
+}