+++ src/models/CmsLang.php
@@ -23,15 +23,13 @@ use yii\helpers\ArrayHelper;
  * @property integer $updated_by
  * @property integer $created_at
  * @property integer $updated_at
- * @property string $active
- * @property string $def
+ * @property boolean $is_active
+ * @property boolean $is_default
  * @property integer $priority
  * @property string $code
  * @property string $name
  * @property string $description
  * @property integer $image_id
- *
- * @property bool $is_default
  *
  * @property CmsSite[] $cmsSites
  * @property CmsStorageFile $image
@@ -73,12 +71,8 @@ class CmsLang extends Core
     public function afterBeforeChecks(Event $e)
     {
         //Если этот элемент по умолчанию выбран, то все остальны нужно сбросить.
-        if ($this->active != Cms::BOOL_Y) {
-            $active = static::find()->where(['!=', 'id', $this->id])->active()->one();
-
-            if (!$active) {
-                $this->active = Cms::BOOL_Y;
-            }
+        if ($this->is_default) {
+            static::updateAll(['is_default' => false]);
         }
     }
 
@@ -88,13 +82,8 @@ class CmsLang extends Core
     public function attributeLabels()
     {
         return array_merge(parent::attributeLabels(), [
-            'id' => Yii::t('skeeks/cms', 'ID'),
-            'created_by' => Yii::t('skeeks/cms', 'Created By'),
-            'updated_by' => Yii::t('skeeks/cms', 'Updated By'),
-            'created_at' => Yii::t('skeeks/cms', 'Created At'),
-            'updated_at' => Yii::t('skeeks/cms', 'Updated At'),
-            'active' => Yii::t('skeeks/cms', 'Active'),
-            'def' => Yii::t('skeeks/cms', 'Default'),
+            'is_active' => Yii::t('skeeks/cms', 'Active'),
+            'is_default' => Yii::t('skeeks/cms', 'Default'),
             'priority' => Yii::t('skeeks/cms', 'Priority'),
             'code' => Yii::t('skeeks/cms', 'Code'),
             'name' => Yii::t('skeeks/cms', 'Name'),
@@ -106,7 +95,7 @@ class CmsLang extends Core
     public function attributeHints()
     {
         return array_merge(parent::attributeLabels(), [
-            'active' => \Yii::t('skeeks/cms', 'On the site must be included at least one language'),
+            'is_active' => \Yii::t('skeeks/cms', 'On the site must be included at least one language'),
         ]);
     }
 
@@ -120,14 +109,17 @@ class CmsLang extends Core
             [['created_by', 'updated_by', 'created_at', 'updated_at', 'priority'], 'integer'],
             [['code', 'name'], 'required'],
             [['code'], 'validateCode'],
-            [['active', 'def'], 'string', 'max' => 1],
             [['code'], 'string', 'max' => 5],
             [['name', 'description'], 'string', 'max' => 255],
             [['code'], 'unique'],
             ['priority', 'default', 'value' => 500],
-            ['active', 'default', 'value' => Cms::BOOL_Y],
-            ['def', 'default', 'value' => Cms::BOOL_N],
             [['image_id'], 'safe'],
+
+            ['is_default', 'default', 'value' => false],
+            ['is_default', 'boolean'],
+            
+            ['is_active', 'default', 'value' => true],
+            ['is_default', 'boolean'],
 
             [
                 ['image_id'],
@@ -166,11 +158,22 @@ class CmsLang extends Core
         return $this->hasOne(CmsStorageFile::className(), ['id' => 'image_id']);
     }
 
+
     /**
-     * @return bool
+     * @deprecated
+     * @return string
      */
-    public function getIs_default()
+    public function getDef()
     {
-        return (bool) ($this->def == 'Y');
+        return $this->is_default ? "Y" : "N";
+    }
+
+    /**
+     * @deprecated
+     * @return string
+     */
+    public function getActive()
+    {
+        return $this->is_active ? "Y" : "N";
     }
 }